发布新版本

+++ socialuni-server/socialuni-official-web/src/main/java/com/socialuni/embed/demo/controller/TestUserRepository.java
@@ -3,9 +3,9 @@ package com.socialuni.embed.demo.controller;
 import com.socialuni.embed.demo.model.UserDO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestUserRepository extends JpaRepository<UserDO, Integer> {
+public interface TestUserRepository extends JpaRepository<UserDO, Long> {
 
     UserDO findByName(String name);
 
-    UserDO findOneById(Integer id);
+    UserDO findOneById(Long id);
 }
